Implement wlr_server_decoration with bare manager
This adds an interface to wlr/types/wlr_server_decoration.h. Only the
core part, the server decoration manager, is added, as this is sufficent
for most compositor use-cases to allow for clients using this protocol
to detect whether they should be using server or client-side
decorations. This wlroots interface is becoming deprecated in a future
version of wlroots (not necessarily the next version). However, this
will not happen until GTK, which uses this protocol, updates to the
alternative xdg_decoration. Until these updates have been done, this
protocol is useful to keep window decorations consistent.